refactor: Divide classes into packages
