+++ src/main/java/chiikawa/ChiikawaException.java
@@ -1,3 +1,5 @@
+package chiikawa;
+
 public class ChiikawaException extends Exception {
     public ChiikawaException(String message) {
         super(message);
