+++ src/main/java/chiikawa/Storage.java
@@ -1,8 +1,11 @@
+package chiikawa;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
+import chiikawa.task.*;
 
 public class Storage {
     private String filePath;
