+++ src/main/java/chiikawa/command/Command.java
@@ -1,3 +1,10 @@
+package chiikawa.command;
+
+import chiikawa.Ui;
+import chiikawa.TaskList;
+import chiikawa.Storage;
+import chiikawa.ChiikawaException;
+
 public abstract class Command {
     protected boolean isExit = false;
