+++ src/main/java/chiikawa/command/UnmarkCommand.java
@@ -1,15 +1,21 @@
-import java.util.Objects;
+package chiikawa.command;
 
-public class MarkCommand extends Command {
+import chiikawa.Ui;
+import chiikawa.TaskList;
+import chiikawa.Storage;
+import chiikawa.ChiikawaException;
+import chiikawa.task.Task;
+
+public class UnmarkCommand extends Command {
     private String indexStr;
 
-    public MarkCommand(String command) {
+    public UnmarkCommand(String command) {
         this.indexStr = command;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws ChiikawaException {
-        if (Objects.equals(this.indexStr, "")) {
+        if (this.indexStr.isEmpty()) {
             throw new ChiikawaException("you give me nothin!! delete wat?!");
         }
 
@@ -20,12 +26,10 @@ public class MarkCommand extends Command {
                 throw new ChiikawaException("no more, wat u doin!!");
             }
 
-            Task markedTask = tasks.markTask(index);
-            ui.showMark(markedTask);
+            Task unmarkedTask = tasks.unmarkTask(index);
+            ui.showUnmark(unmarkedTask);
         } catch (NumberFormatException e) {
             throw new ChiikawaException("giv 1 numba!! 1!! number!!! only!!!!");
         }
-
-
     }
 }
