+++ src/main/java/chiikawa/task/Task.java
@@ -1,3 +1,5 @@
+package chiikawa.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.regex.Matcher;
